Update and rename mealPlanner.php to MealPlanner.php
La classe "MealPlanner" est une "sous-classe" de la classe "Client"
