--- src/back/classes/business/model/mealPlanner.php
@@ -1,143 +1,0 @@
-MealPlanner
-
-/*
-<?php
-
-namespace classes\business\model;
-
-use classes\AutoLoader;
-AutoLoader::register();
-
-class Client
-{
-    private int $id;
-    private string $first_name;
-    private string $last_name;
-    private string $civility;
-    private string $pseudo;
-    private string $mail;
-    private string $password;
-    private array $assessed_recipes;
-    private array $recorded_recipes;
-    private array $preferences_ingredients;
-
-    /**
-     * Client constructor.
-     */
-    public function __construct(int $id, string $first_name, string $last_name, string $civility, string $pseudo,
-                                string $mail, string $password, array $assessed_recipes = array(),
-                                array $recorded_recipes = array(), array $preferences_ingredients = array())
-    {
-        $this->id = $id;
-        $this->first_name = $first_name;
-        $this->last_name = $last_name;
-        $this->civility = $civility;
-        $this->pseudo = $pseudo;
-        $this->mail = $mail;
-        $this->password = $password;
-        $this->assessed_recipes = $assessed_recipes;
-        $this->recorded_recipes = $recorded_recipes;
-        $this->preferences_ingredients = $preferences_ingredients;
-    }
-
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    public function setId(int $id)
-    {
-        $this->id = $id;
-    }
-
-    public function getFirstName(): string
-    {
-        return $this->first_name;
-    }
-
-    public function setFirstName(string $first_name)
-    {
-        $this->first_name = $first_name;
-    }
-
-    public function getLastName(): string
-    {
-        return $this->last_name;
-    }
-
-    public function setLastName(string $last_name)
-    {
-        $this->last_name = $last_name;
-    }
-
-    public function getCivility(): string
-    {
-        return $this->civility;
-    }
-
-    public function setCivility(string $civility)
-    {
-        $this->civility = $civility;
-    }
-
-    public function getPseudo(): string
-    {
-        return $this->pseudo;
-    }
-
-    public function setPseudo(string $pseudo)
-    {
-        $this->pseudo = $pseudo;
-    }
-
-    public function getMail(): string
-    {
-        return $this->mail;
-    }
-
-    public function setMail(string $mail)
-    {
-        $this->mail = $mail;
-    }
-
-    public function getPassword(): string
-    {
-        return $this->password;
-    }
-
-    public function setPassword(string $password)
-    {
-        $this->password = $password;
-    }
-
-    public function getAssessedRecipes(): array
-    {
-        return $this->assessed_recipes;
-    }
-
-    public function setAssessedRecipes(array $assessed_recipes)
-    {
-        $this->assessed_recipes = $assessed_recipes;
-    }
-
-    public function getRecordedRecipes(): array
-    {
-        return $this->recorded_recipes;
-    }
-
-    public function setRecordedRecipes(array $recorded_recipes)
-    {
-        $this->recorded_recipes = $recorded_recipes;
-    }
-
-    public function getPreferencesIngredients(): array
-    {
-        return $this->preferences_ingredients;
-    }
-
-    public function setPreferencesCategories(array $preferences_ingredients)
-    {
-        $this->preferences_ingredients = $preferences_ingredients;
-    }
-}
-*/
